Fix field sizes for object header

+++ implementation/src/hdf5/object_header.h
@@ -22,10 +22,10 @@ struct ObjectHeader {
     // includes continuation messages
     uint16_t message_count{};
     // number of hard links to this object in the current file
-    uint64_t object_ref_count{};
+    uint32_t object_ref_count{};
     // number of bytes of header message data for this header
     // does not include size of object header continuation blocks
-    uint64_t object_header_size{};
+    uint32_t object_header_size{};
     // messages
     std::vector<ObjectHeaderMessage> messages{};
